requirements.txt
httpx
notion-client
pytest
pytest-env
pytest-cov
coveralls
feedparser
tradernet-sdk
